% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{add_release_url}
\alias{add_release_url}
\title{Add a release URL to the registry}
\usage{
add_release_url(date_utc, title, url)
}
\arguments{
\item{date_utc}{Character. e.g. "March 02 2026 09:00 UTC"}

\item{title}{Character. e.g. "HCOB Eurozone Manufacturing PMI"}

\item{url}{Character. Full URL to the press release.}
}
\description{
Add a release URL to the registry
}
