% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{filter_releases}
\alias{filter_releases}
\title{Filter release list for relevant releases}
\usage{
filter_releases(releases, region = "Eurozone", type = "all")
}
\arguments{
\item{releases}{Data frame from get_release_list()}

\item{region}{Character. "Eurozone" or "Germany"}

\item{type}{Character. "flash", "final_composite", "final_manufacturing", or "all"}
}
\value{
Filtered data frame
}
\description{
Filter release list for relevant releases
}
