% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_release_registry}
\alias{get_release_registry}
\title{Get all registered release URLs}
\usage{
get_release_registry()
}
\value{
Data frame with columns: date_utc, title, url
}
\description{
Get all registered release URLs
}
