% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pmi_config}
\alias{pmi_config}
\title{Define PMI table/series configuration}
\usage{
pmi_config()
}
\value{
List of table configurations
}
\description{
Each region gets one table with 4 series:
- composite: Composite PMI Output Index
- services: Services PMI Business Activity Index
- manufacturing: Manufacturing PMI (headline)
- manufacturing_output: Manufacturing PMI Output Index
}
\details{
Release structure:
- Eurozone/Germany: flash (~20th) + final manufacturing (~1st) +
  final composite/services (~3rd-4th)
- Italy/France/Spain: no flash, manufacturing (~1st) + services (~3rd-4th)
}
