% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_series_table}
\alias{prepare_series_table}
\title{Prepare rows for the `series` table}
\usage{
prepare_series_table(config_name, con, schema = "platform")
}
\arguments{
\item{config_name}{character}

\item{con}{database connection}

\item{schema}{database schema, default "platform"}
}
\value{
data frame ready for UMARimportR::insert_new_series()
  (one row per series)
}
\description{
Creates 4 series per region table (one per indicator).
}
