% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_dimensions_table}
\alias{prepare_table_dimensions_table}
\title{Prepare row for the `table_dimensions` table}
\usage{
prepare_table_dimensions_table(config_name, con, schema = "platform")
}
\arguments{
\item{config_name}{character}

\item{con}{database connection}

\item{schema}{database schema, default "platform"}
}
\value{
data frame ready for UMARimportR::insert_new_table_dimensions()
}
\description{
PMI tables have a single non-time dimension: "indicator"
}
