% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_table}
\alias{prepare_table_table}
\title{Prepare row for the `table` table}
\usage{
prepare_table_table(config_name, source_id, con, schema = "platform")
}
\arguments{
\item{config_name}{character, one of names(pmi_config()): "eurozone",
"germany", "france", "italy", "spain"}

\item{source_id}{probably 9}

\item{con}{database connection}

\item{schema}{database schema, default "platform"}
}
\value{
data frame ready for UMARimportR::insert_new_table_table()
}
\description{
Prepare row for the `table` table
}
